Improved the LED tracking and velocity generation
The velocity is not stored as a table anymore but as a piecewise linear
function of the table. This will allow the use on the micro-controller
without wasting space for the table.

The LED tracking now only operates on every second event. In addition,
the weight computation is improved and only relevant events will be
incorporated. Note that every 10th event has the chance to get used even
if it is deemed irrelevant to allow the tracker to re-latch onto the LED
in case it got lost.